Color based sorting of objects from conveyer belt
This is a multi-disciplinary project including electronics along with
mechanical skills. We made a small mechanical structure of a conveyor
belt. It also involves processing on live video input and then sends
command to controller. This project was used to sort two different
colored objects moving on the conveyor. When one of the two colors are
detected by the camera, command is sent to controller to stop the
rotation of the belt,  pick up the object  and place it on the
appropriate place.